Rename file from component to extension
