+++ packages/esm-patient-registration-app/src/index.ts
@@ -3,7 +3,7 @@ import { esmPatientRegistrationSchema } from './config-schema';
 import { moduleName, patientRegistration } from './constants';
 import { setupOffline } from './offline';
 import rootComponent from './root.component';
-import addPatientLinkComponent from './add-patient-link.component';
+import addPatientLinkComponent from './add-patient-link.extension';
 import editPatientDetailsButtonComponent from './widgets/edit-patient-details-button.component';
 import { PatientPhotoExtension } from './patient-photo.extension';
